Adds setuputils support
Benefits of setuputils: Easy distribution / install via PyPI; easier
for distributions to package, easier to support Windows and MacOSX
users.

Required restructuring pcbmode into a module, rather than a
script. Also moves utils module to a submodule of pcbmode. This is
helpful should others want to use pcbmode as a library (e.g. from
pcbmode.utils.svg import blah).

Looks for pcbmode_config.json in the invocation directory first, falls
back to the default (which is stored with the module in the install
directory).