Update setup instructions
The current state of the code requires the code to be installed as a
package before using it. This commit updates the setup instructions so
that they reflect the current state of the code.

This change is based on the work in progress by user nocko on
GitHub. I reused the parts relating to building from source with
virtualenv. The other methods (installing binary packages) are not
available yet, so I skipped them.

Changes:

- Replace "python PCBmodeE/pcbmode.py" by "pcbmode"
- New section on Setup page: "Installation from Source with virtualenv"